fix(FEC-11639): webcast Q&A icons multiplied (#276)
**the issue:**
we have 2 use-cases that can cause multiple Q&A icons:
1. failover primary to backup and vice-versa - **working as expected**:
as a result of player.reset, unloadMedia and loadMedia are working as expected and contib is cleaning the icons.
2. broadcast is offline (both streams are down), executing attachMedia and detachMedia - **having multiple icons**:
player.reset is not part of the flow in this use-case.
detachMedia is being executed, which is not calling unloadMedia. Therefore, icon is not being cleaned. Then when loadMedia is triggered, another icon is being added to player.

**the solution:**
resetting kitchenSinkItem obj in unloadMedia.
checking if there is a kitchenSinkItem that already exists, before creating new one.

**Note:**
The current fix _will_ fix the issue for 2nd use-case, however, root cause is not really solved.
Root issue is that contrib is doing the cleaning as a result of player.reset, which is only triggered for the 1st use-case. In addition, in the 2nd use-case player.reset is not triggered, but loadMedia is, which causes the multiple icons.

Solves [FEC-11639](https://kaltura.atlassian.net/browse/FEC-11639)